[Decomp][Unit] Cohesion split: extract melee damage calc/apply to UnitMeleeDamage.cpp
